timetable/pyproject.toml
[tool.poetry]
name = "timetable"
version = "0.1.0"
description = ""
authors = ["ligets <legen2a208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
pyjwt = "^2.9.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.35"
pydantic = "^2.9.2"
alembic = "^1.13.3"
httpx = "^0.27.2"
datetime = "^5.5"
pydantic-settings = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
